requirements.txt
emoji==2.4.0
gspread==5.8.0
numpy==1.20.3
oauth2client==4.1.3
pandas==1.4.3
pandas==1.4.1
python-telegram-bot==13.1
pytz==2021.3
